vaev-layout: Split base.cpp into multiple files.
