+++ src/vaev-engine/layout/paint.cpp
@@ -1,16 +1,20 @@
 export module Vaev.Engine:layout.paint;
 
-import Karm.Image;
-import Karm.Gc;
-import Karm.Scene;
+import Karm.Core;
 import Karm.Debug;
+import Karm.Gc;
 import Karm.Gfx;
-import Karm.Math;
+import Karm.Image;
 import Karm.Logger;
+import Karm.Math;
+import Karm.Scene;
+
 
 import :style;
+import :values;
 import :layout.base;
 import :layout.values;
+import :dom.node;
 
 namespace Vaev::Layout {
