+++ src/vaev-engine/layout/writing.cpp
@@ -3,8 +3,7 @@ module;
 export module Vaev.Engine:layout.writing;
 
 import :values;
-
-import :layout.base;
+import :layout.box;
 
 namespace Vaev::Layout {
